test: run the runtime probe for every test job
Move _cgroup/_env_probe from tests/appsec/contrib_appsec/conftest.py to the
repo-root conftest.py, the local pytest plugin already loaded for every
pytest run from the repo root. This reaches every test job in the GitLab
pipeline without touching .gitlab config or registering a new plugin.

Two changes were needed to make it safe repo-wide:

- xdist guard: 13 suites already run -n auto, and pytest_configure fires in
  every worker, so an unguarded probe would print once per worker and rerun
  the benchmarks each time. _probe_enabled() returns False when config has
  workerinput, so only the controller emits.
- gating: the probe previously ran unconditionally, which is fine for one
  suite but noisy for every local run of every suite. It now runs when
  CI=true, which is what covers the pipeline, or DD_TEST_RUNTIME_PROBE=1.

scripts/ddtest forwards only DD_TRACE_AGENT_URL on the docker path (the
TEST_* passthrough is podman only), so the local override needs its own -e
line or it silently does nothing.

Measured cost 0.55s per job. Verified: exactly one start and one end probe
per session with 6 xdist workers active, none when ungated.
