+++ scripts/ddtest
@@ -126,6 +126,7 @@ if [ "${DD_TEST_PODMAN_COMPAT:-0}" = "1" ]; then
 else
     docker compose ${COMPOSE_FILES[@]} run $BUILD_FLAG \
            -e DD_TRACE_AGENT_URL \
+           -e DD_TEST_RUNTIME_PROBE \
            --rm \
            -i \
            testrunner \
